fix forward pass

+++ train_gpt.py
@@ -2498,7 +2498,7 @@ class GPT(nn.Module):
                 for s in range(n_rep):
                     if n_rep > 1 and len(self.intra_loop_controllers) > 0:
                         ctrl = self.intra_loop_controllers[i - self.intra_loop_start]
-                        out = ctrl(x.mean(dim=1)).view(x.shape[0], self.intra_loop_steps, 2, self._intra_model_dim)
+                        out = ctrl(x.mean(dim=1).float()).view(x.shape[0], self.intra_loop_steps, 2, self._intra_model_dim)
                         scale = out[:, s, 0, :].unsqueeze(1).to(dtype=x.dtype)  # [B,1,dim]
                         shift = out[:, s, 1, :].unsqueeze(1).to(dtype=x.dtype)  # [B,1,dim]
                         x = x * (1.0 + scale.tanh()) + shift
@@ -2512,7 +2512,7 @@ class GPT(nn.Module):
                 for s in range(n_rep):
                     if n_rep > 1 and len(self.intra_loop_controllers) > 0:
                         ctrl = self.intra_loop_controllers[j - self.intra_loop_start]
-                        out = ctrl(x.mean(dim=1)).view(x.shape[0], self.intra_loop_steps, 2, self._intra_model_dim)
+                        out = ctrl(x.mean(dim=1).float()).view(x.shape[0], self.intra_loop_steps, 2, self._intra_model_dim)
                         scale = out[:, s, 0, :].unsqueeze(1).to(dtype=x.dtype)
                         shift = out[:, s, 1, :].unsqueeze(1).to(dtype=x.dtype)
                         x = x * (1.0 + scale.tanh()) + shift
@@ -2568,7 +2568,7 @@ class GPT(nn.Module):
                 for s in range(n_rep):
                     if n_rep > 1 and len(self.intra_loop_controllers) > 0:
                         ctrl = self.intra_loop_controllers[i - self.intra_loop_start]
-                        out = ctrl(x.mean(dim=1)).view(x.shape[0], self.intra_loop_steps, 2, self._intra_model_dim)
+                        out = ctrl(x.mean(dim=1).float()).view(x.shape[0], self.intra_loop_steps, 2, self._intra_model_dim)
                         scale = out[:, s, 0, :].unsqueeze(1).to(dtype=x.dtype)
                         shift = out[:, s, 1, :].unsqueeze(1).to(dtype=x.dtype)
                         x = x * (1.0 + scale.tanh()) + shift
@@ -2583,7 +2583,7 @@ class GPT(nn.Module):
                 for s in range(n_rep):
                     if n_rep > 1 and len(self.intra_loop_controllers) > 0:
                         ctrl = self.intra_loop_controllers[j - self.intra_loop_start]
-                        out = ctrl(x.mean(dim=1)).view(x.shape[0], self.intra_loop_steps, 2, self._intra_model_dim)
+                        out = ctrl(x.mean(dim=1).float()).view(x.shape[0], self.intra_loop_steps, 2, self._intra_model_dim)
                         scale = out[:, s, 0, :].unsqueeze(1).to(dtype=x.dtype)
                         shift = out[:, s, 1, :].unsqueeze(1).to(dtype=x.dtype)
                         x = x * (1.0 + scale.tanh()) + shift
